# specs/002-coach-availability-impersonation/contracts/openapi.yaml
openapi: 3.1.0
info:
  title: PracticePerfect — Coach Invitations, Availability & Impersonation
  version: 1.3.0
  description: |
    Contract for feature 002-coach-availability-impersonation. It is version **1.3.0** of the same
    `/api/v1` contract feature 001 defined at 1.2.0, and it is **purely additive**: every operation
    below is new, and the only changes to existing response models are new optional fields
    (`CurrentUser.impersonation`, `CurrentUser.impersonation_ended`,
    `TrainerPlayerSummary.availability`, `TrainerPlayerSummary.availability_updated_at`). No
    operation is removed, and no existing field changes type or meaning, so the frontend compiled
    against 1.2.0 keeps working until it chooses to read the new fields (research.md R2-22).

    This file documents **only what 1.3.0 adds**. Operations defined in
    `specs/001-user-roles-admin/contracts/openapi.yaml` are unchanged and not repeated here, with
    the single exception of the two schema extensions above, which are given in full so the
    frontend types have one authoritative source.

    Three groups of operations:

    * **Coach invitations** (`coach-invitations` tag, plus `/trainer/coach-invitations` and
      `/trainer/coaches`) — a trainer invites a coach, tracks the invitation, and manages the
      roster; the invited person previews, registers, or accepts. US-01.08, FR-001 – FR-023.
    * **Availability** (`availability` tag) — a coach's own week, a player profile's week, and the
...

...

...
servers:
...
security:
  - sessionCookie: []
paths:
...
  /me/availability:
    get:
      tags: [availability]
      summary: Read the signed-in coach's own stated week
      operationId: getOwnAvailability
      x-requirements: [FR-024, FR-032, FR-033, FR-035]
      description: |
        Coach only. An account that has never stated times returns `slots: []` with
        `updated_at: null` — "no times set", which FR-035 requires be distinguishable from
        unavailable and from a week that was deliberately cleared (`slots: []` with a non-null
        `updated_at`).
      responses:
        '200':
          description: The coach's stated week.
          content:
            application/json:
              schema: { $ref: '#/components/schemas/AvailabilityWeek' }
        '401': { $ref: '#/components/responses/Unauthenticated' }
        '403': { $ref: '#/components/responses/Forbidden' }

    put:
      tags: [availability]
      summary: Replace the signed-in coach's stated week
      operationId: replaceOwnAvailability
      x-requirements: [FR-024, FR-026, FR-027, FR-028, FR-029, FR-032, FR-033]
      description: |
        Replaces the whole week in one transaction: the submitted set is validated in full before
        any row is touched, so a refused save leaves the previous week exactly as it was and no week
        is ever observable half-saved (FR-027, FR-029). Two concurrent saves resolve as
        last-complete-save-wins; there is no version token to send.

        `422` carries the offending **day** and the reason — overlapping ranges, a start at or after
        its end, a range past the end of the day, or more than six ranges in a day.
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/AvailabilityWeekUpdate' }
      responses:
        '200':
          description: The stored week, as saved.
          content:
            application/json:
              schema: { $ref: '#/components/schemas/AvailabilityWeek' }
        '401': { $ref: '#/components/responses/Unauthenticated' }
        '403': { $ref: '#/components/responses/Forbidden' }
        '422': { $ref: '#/components/responses/ValidationError' }

    delete:
      tags: [availability]
      summary: Clear the signed-in coach's stated week
      operationId: clearOwnAvailability
      x-requirements: [FR-030, FR-032]
      description: |
        Equivalent to saving an empty week, and kept as its own operation because "I have no times
        to offer" is a statement a person makes deliberately. `updated_at` is stamped, so the week
        reads as cleared-on-a-date rather than never-stated.
      responses:
        '204': { description: The week is cleared. }
        '401': { $ref: '#/components/responses/Unauthenticated' }
        '403': { $ref: '#/components/responses/Forbidden' }
...
components:
  securitySchemes:
    sessionCookie:
      type: apiKey
      in: cookie
      name: pp_session
      description: |
        Unchanged from feature 001. While an impersonation is live the same cookie resolves to the
        impersonated account; no second credential is issued (research.md R2-14).
...
  responses:
    Unauthenticated:
      description: No usable session cookie.
      content:
        application/json:
          schema: { $ref: '#/components/schemas/Error' }
    Forbidden:
      description: The caller's role or ownership does not permit this action.
      content:
        application/json:
          schema: { $ref: '#/components/schemas/Error' }
...
    ValidationError:
      description: One or more fields are invalid; each offending field is identified.
      content:
        application/json:
          schema: { $ref: '#/components/schemas/ValidationErrorResponse' }
  schemas:
...
    Error:
      type: object
      additionalProperties: false
      required: [error]
      properties:
        error:
          type: object
          additionalProperties: true
          required: [code, message]
          properties:
            code: { type: string }
            message: { type: string }
    ValidationErrorResponse:
      type: object
      additionalProperties: false
      required: [error]
      properties:
        error:
          type: object
          additionalProperties: false
          required: [message, fields]
          properties:
            message: { type: string }
            fields:
              type: array
              items:
                type: object
                additionalProperties: false
                required: [field, message]
                properties:
                  field: { type: string }
                  message: { type: string }

    # --- Coach invitations ---
...
    AvailabilitySlot:
      type: object
      additionalProperties: false
      required: [day_of_week, start_minute, end_minute]
      description: |
        One range on one day. Minutes from midnight on a 15-minute grid (research.md R2-08):
        `17:00–20:00` is `{day_of_week: 0, start_minute: 1020, end_minute: 1200}`.
      properties:
        day_of_week:
          type: integer
          minimum: 0
          maximum: 6
          description: 0 = Monday … 6 = Sunday.
        start_minute:
          type: integer
          minimum: 0
          maximum: 1425
          multipleOf: 15
        end_minute:
          type: integer
          minimum: 15
          maximum: 1440
          multipleOf: 15
          description: May be 1440 (midnight). Must be greater than `start_minute`.
    AvailabilityWeek:
      type: object
      additionalProperties: false
      required: [slots, updated_at]
      properties:
        slots:
          type: array
          maxItems: 42
          description: |
            Ordered by `(day_of_week, start_minute)`. At most six per day (FR-028), so at most 42.
          items: { $ref: '#/components/schemas/AvailabilitySlot' }
        updated_at:
          type: [string, 'null']
          format: date-time
          description: |
            `null` means never stated. A non-null value with an empty `slots` means deliberately
            cleared. Neither means "unavailable" (FR-035).
    AvailabilityWeekUpdate:
      type: object
      additionalProperties: false
      required: [slots]
      description: |
        The complete week. This is a replacement, not a patch: whatever is absent is removed
        (FR-029). An empty array is equivalent to `DELETE`.
      properties:
        slots:
          type: array
          maxItems: 42
          items: { $ref: '#/components/schemas/AvailabilitySlot' }

    # --- Impersonation ---
...